14_medical_consultation_system/requirements.txt
pandas
spacy
py2neo==2021.2.4
transformers
pytorch-crf
peft
sentencepiece
bitsandbytes